seperate browser-specific code into own modules
