+++ src/test/logic/onAppMount.test.js
@@ -12,110 +12,6 @@ global.chrome = {
 }
 
 describe('onAppMount', () => {
-  describe('browserFirefox', () => {
-    describe('urlChecking', () => {
-      test('a valid url should return string "valid URL"', () => {    
-        const result = onAppMount.browserFirefox.urlChecking(mockTab)
-    
-        expect(result).toEqual('valid URL');
-      })
-    
-      test('an addons url should return addons error text', () => {
-        const tab = {url: 'https://addons.mozilla.org/en-US/firefox/'}
-        
-        const result = onAppMount.browserFirefox.urlChecking(tab)
-    
-        expect(result).toEqual(onAppMount.browserFirefox.addonsErrorString);
-      })
-    })
-  })
-
-  describe('browserChrome', () => {
-    describe('chromeInfo', () => {
-      let userAgentGetter;
-      
-      beforeEach(() => {
-        // navigator object cannot be directly modified so we mock it here
-        userAgentGetter = jest.spyOn(window.navigator, 'userAgent', 'get')
-      })
-
-      test('it returns a truthy string when navigator.userAgent contains the word "Chrome"', () => {
-        userAgentGetter.mockReturnValue('Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.135 Safari/537.36 OPR/70.0.3728.154')
-
-        const result = onAppMount.browserChrome.chromeInfo();
-
-        expect(result).toBeTruthy();
-      })
-
-      test('it returns null when navigator.userAgent does not contain the word "Chrome"', () => {
-        userAgentGetter.mockReturnValue('Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:61.0) Gecko/20100101 Firefox/61.0');
-
-        const result = onAppMount.browserChrome.chromeInfo();
-
-        expect(result).toBeFalsy();
-      })
-    })
-
-    describe('urlChecking', () => {
-      test('a valid url should return string "valid URL"', () => {
-        const result = onAppMount.browserChrome.urlChecking(mockTab)
-    
-        expect(result).toEqual('valid URL');
-      })
-    
-      test('a webstore url should return webstore error text', () => {
-        const tab = {url: 'https://chrome.google.com/webstore/category/extensions'}
-        
-        const result = onAppMount.browserChrome.urlChecking(tab)
-    
-        expect(result).toEqual(onAppMount.browserChrome.webstoreErrorString);
-      })
-    })
-  })
-
-  describe('browserOpera', () => {
-    describe('operaInfo', () => {
-      let userAgentGetter;
-      
-      beforeEach(() => {
-        // navigator object cannot be directly modified so we mock it here
-        userAgentGetter = jest.spyOn(window.navigator, 'userAgent', 'get')
-      })
-
-      test('it returns a truthy string when navigator.userAgent contains the word "Opera" or the string "OPR"', () => {
-        userAgentGetter.mockReturnValue('Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.135 Safari/537.36 OPR/70.0.3728.154')
-
-        const result = onAppMount.browserOpera.operaInfo();
-
-        expect(result).toBeTruthy();
-      })
-
-      test('it returns null when navigator.userAgent does not contain the word "Opera" or the string "OPR"', () => {
-        userAgentGetter.mockReturnValue('Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:61.0) Gecko/20100101 Firefox/61.0');
-
-        const result = onAppMount.browserOpera.operaInfo();
-
-        expect(result).toBeFalsy();
-      })
-    })
-
-    describe('urlChecking', () => {
-      test('a valid url should return string "valid URL"', () => {
-        const result = onAppMount.browserOpera.urlChecking(mockTab)
-    
-        expect(result).toEqual('valid URL');
-      })
-    
-      test('a opera addons url should return addons error text', () => {
-        const tab = {url: 'https://addons.opera.com'}
-        
-        const result = onAppMount.browserOpera.urlChecking(tab)
-    
-        expect(result).toEqual(onAppMount.browserOpera.addonsErrorString);
-      })
-    })
-  })
-
   describe('userBrowser', () => {
     test('it correctly identifies Firefox', () => {
       const expected = 'firefox';
